Refactor sensor and relay handling, update configuration settings

+++ Config/FanControlSettings.cs
@@ -4,5 +4,5 @@ public class FanControlSettings
 {
     public double HumidityThreshold { get; set; }
     public int RelayGpioPin { get; set; }
-    public double ReadIntervalMs { get; set; }
+    public int ReadIntervalMs { get; set; }
 }
